+++ ShipmentGeek/Program.cs
@@ -16,7 +16,9 @@ namespace ShipmentGeek
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new frmMain());
+            Application.Run(MainForm = new frmMain());
         }
+
+        public static frmMain MainForm;
     }
 }